+++ filters/randvelocity.c
@@ -29,13 +29,13 @@ filter_midi_randvelocity(MidiFilter* self,
 		const uint8_t* const buffer,
 		uint32_t size)
 {
-	const int chs = midi_limit_chn(floor(*self->cfg[0]) -1);
+	const int chs = midi_limit_chn(floorf(*self->cfg[0]) -1);
 	const int chn = buffer[0] & 0x0f;
 	int mst = buffer[0] & 0xf0;
 
 	if (size != 3
 			|| !(mst == MIDI_NOTEON || (mst == MIDI_NOTEOFF))
-			|| !(floor(*self->cfg[0]) == 0 || chs == chn)
+			|| !(floorf(*self->cfg[0]) == 0 || chs == chn)
 			)
 	{
 		forge_midimessage(self, tme, buffer, size);
